stan: math: Include master headers instead of individual files
The includes of Stan Math should be changed to include master headers instead of individual files. The only exception is the Eigen.hpp header file. That one must be included first because we override internal Eigen stuff.

This would make the interface between `rstan` and Stan Math more stable and less prone to issues when some files are moved in Stan Math.

Co-authored-by: Rok Češnovar <rok.cesnovar@fri.uni-lj.si>