Add "file://" protocol to stan_progress.html
Adding the "file://" protocol to the beginning of the iframe increases compatibility with other browsers. Before, the update page would work in R Studio (which is why I think it hasn't been fixed so far) and using Edge/IE. Firefox complained about an invalid protocol, and would not display the page. After this simple change, it works as expected, and eliminates the need for "open_progress = FALSE" when using something else, like Emacs and ESS.